requirements.txt
crewai>=0.28.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
